nfo added
title:meh Thouht could add descipt below in box but this will his the revision  i think so -- well title this added nfo

Dynamically set system enviroment variables for use in scripts or whatever. 
If no input is received it'll ask you what you want. 
*No checks beyond input 1 and input 2, so if you skip past the prompt it'll run what it has (might change that)
Persist an array outside of all setlocal, setup an enviroment for some command line app, change a variable to something else on the fly, whatever you need really. 

All is callable, always, from the prompt until it's removed. (Might work on finding out how this can be done [which I'm sure it can be] )
As was stated in the sauce: "it's akin to going to system settings / advanced/ enviroment variables"